+++ loader/app/db/models.py
@@ -1,3 +1,7 @@
+# TODO at the moment, this is mostly a copy of backend's models.
+# Once the migration is done and the code works, and is tested, then put the common part
+# of the model in the common module.
+
 import enum
 
 import sqlalchemy as sa
@@ -79,7 +83,7 @@ class Document(Auditable, Base):
     loaded_ts = sa.Column(sa.DateTime(timezone=True), onupdate=func.now())
     name = sa.Column(sa.Text, nullable=False)
     source_url = sa.Column(sa.Text)
-    source_id = sa.Column(sa.BigInteger, sa.ForeignKey(Source.id), nullable=False)
+    source_id = sa.Column(sa.Integer, sa.ForeignKey(Source.id), nullable=False)
     url = sa.Column(sa.Text)
 
     # this will be in the loader DB
@@ -99,7 +103,7 @@ class Sector(Base):  # noqa: D101
     parent_id = sa.Column(sa.Integer, sa.ForeignKey("sector.id"))
     name = sa.Column(sa.Text, nullable=False)
     description = sa.Column(sa.Text, nullable=False)
-    source_id = sa.Column(sa.BigInteger, sa.ForeignKey(Source.id), nullable=False)
+    source_id = sa.Column(sa.Integer, sa.ForeignKey(Source.id), nullable=False)
 
 
 class DocumentSector(Base):  # noqa: D101
@@ -117,7 +121,7 @@ class Instrument(Base):  # noqa: D101
     parent_id = sa.Column(sa.Integer, sa.ForeignKey("instrument.id"))
     name = sa.Column(sa.Text, nullable=False)
     description = sa.Column(sa.Text, nullable=False)
-    source_id = sa.Column(sa.BigInteger, sa.ForeignKey(Source.id), nullable=False)
+    source_id = sa.Column(sa.Integer, sa.ForeignKey(Source.id), nullable=False)
 
 
 class DocumentInstrument(Base):  # noqa: D101
